+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Added an example `example_pair_object_form.py`.
 * Added `absolute_height` option to the `TextObject`.
 * Added `font` option to `TextObject`.
 * Multi-cursor visual effects.